feat(robot-server): add a new endpoint to upload csv data files (#15481)
Closes AUTH-426

# Overview

Adds a new router for uploading and managing data files. The new
endpoint can accept a single data file, either as a file upload or a
path to the file on the robot.

The endpoint saves the uploaded file to the persistent directory on the
robot, adds the file's info to the data_files database, and returns the
file ID and other deets in the response.

Duplicate file uploads use the previously saved file & database entry
instead of creating a new one.

# Test Plan

- On a robot, upload a data file using the `POST /dataFiles` endpoint
and providing a file in the request. Verify that-
- [ ] the file is uploaded to the persistent storage at location-
`../opentrons-robot-server/data_files/<file_id>/`
    - [ ] the file's info is added to the database
    - [ ] the file's ID is returned in the response
- [ ] On a robot, send a file to `POST /dataFiles` by providing a
`file_path` to a data file on the robot (the file should be outside the
persistent directory) and verify all the above
- [ ] Send an already existing file to the server and verify that no new
file is saved and no new entry is added to database. Verify that the
response contains the previously assigned file ID

# Changelog

- added new router, models and data files store for the new endpoint
- added unit & integration tests

# Review requests

- check that the implementation meets client's expectations
- any suggestions on placement of file_reader_writer and file_hasher
classes/functions? They have been so far associated with only protocols
and hence were part of the protocols/ api packages but I think it's time
to move it to a common utility
- any other code suggestions

# Risk assessment

None to the existing system. New, isolated feature.