feat(app, api-client, react-api-client): send CSV RTP file to robot before creating protocol and run (#15527)
The new RTP paradigm for utilizing CSV files as runtime parameters
requires sending the ID of the file to be used as the value in the
`runTimeParameterValues` object sent with `createProtocol` and
`createRun`. Here, we send any selected RTP files to the new
`/dataFiles` endpoint and await the response containing an ID for each
file. We then associate the returned ID with its respective runtime
parameter variable name, and send that key:value pair along with the
other value runtime parameters.