chore(api,robot-server): Upgrade to aionotify 0.3.1 (#15546)
# Overview

The `opentrons` Python package is not yet compatible with Python 3.11.
It raises errors upon import, which forces scientists to use the soon to
be outdated Python 3.10.

The `aionotify` dependency has since been updated for Python 3.11
compatibility, so just a dependency upgrade is needed.

# Changelog

* Update `aionotify` to current version 0.3.1

# Review requests

I'm not familiar with how this repository is organized, or how the
Python package ending up on PyPI is built.

I hope that the CI pipeline turns green, and that maintainers can take
over from there. Hopefully this remains a really easy fix for the
long-standing compatibility issue #12006.