refactor: use deckmap in error recovery
Two places in error recovery, refactoring as we go.

One big change was that there was a bunch of stuff called RecoveryMap as
in recovery (deck) map, but recovery also has the Recovery Map, the big
data structure that holds how all the steps of error recovery flow
together. This is very confusing, so now what used to be
RecoveryMap (component, utils) is now DeckMap.