feat(app): add CommandIndex component
This is a little component that will render a command index on only
desktop. We may want this to work slightly differently but it seems like
a good start. It's worth componentizing because there are already and
probably will be more weird little details around the width when used in
a stack of command objects to make them not all different lengths.