refactor(app): add and use deck map component in interventionmodal (#15570)
Adds DeckMapContent to intervention modal. This is a small wrapper
around one of two different kinds of deck map:
- A labware-rendering map that is also capable of rendering highlights
around some labware, used in some selection screens in error recovery
- A deck config style block map that supports clicking on some of the
blocks to select them, used in some selection screens in the drop tip
wizard (and thus in error recovery)

Stories for the deckmap are [on
storybook](https://s3-us-west-2.amazonaws.com/opentrons-components/exec-501-deckmap-content/index.html?path=/docs/app-molecules-interventionmodal-deckmapcontent--docs),
where it's rendered side-by-side with the standin since it is only
destined to be used in a two-column layout.

Extra fun changes:
- Use the ODD text size args for _all_ slot labels in `BaseDeck`,
because otherwise they are completely unreadable once the deckmap gets
to small. I think this looks a lot better everywhere, and is actually
readable when the deckmap is small, so let's go with it; may want to
come back and make these something more specific on desktop
- Error recovery has a `RecoveryMap`, a large data structure that is
core to the wizard flow and defines how users move between steps and
screens. Error recovery also had a `RecoveryMap`, which was a component
for rendering a deckmap, and a `useRecoveryMapUtils`, a hook for getting
that component's arguments. Now it uses the `DeckMapContent` component
above directly, and `useDeckMapUtils`, a hook for getting that
component's arguments
- There's something about the deck config style map that isn't quite
right and I think that it's because of this:
https://opentrons.atlassian.net/browse/EXEC-513

Review requests:
- i'm pretty sure what I did to drop tip will work but it might be
reasonably called somewhat gross. this is all somewhat gross though, so
c'est la vie
- the deckmap slot labels do not render right in firefox at these sizes.
if you're looking at storybook, you're going to have to use chrome
- fix an issue that was preventing ER wizard from being displayed on ODD
Testing, todo:
- [ ] error recovery
- [ ] drop tip in error recovery
- [ ] drop tip outside of error recovery


Closes EXEC-501