fix(protocol-designer): getPipetteCapacity correctly updates tip capa… (#15595)
closes AUTH-553

This is a bug from PR 15096, in that PR, i changed it so for multi tiprack support, the tiprack key that is required for the transfer and mix forms changed from the tiprack id to the tiprack def uri. But i forgot to migrate it to this util!! This resulted in the user being unable to get accurate tip volume capacity.