fix(protocol-designer): fix logic in safePipetteMovements collision detection
Several errors in `safePipetteMovements` collision logic were causing errors to be raised in
`transfer` command creator calls using a 96-channel pipette with COLUMN configuration. This PR
addresses these errors by 1) passing `wellLocationPoint` to
`getPipetteBoundsAtSpecifiedMoveToPosition` to get correct pipette bounds, 2) passing the
surrounding labware ID rather than the target labware ID to `getHighestZInSlot`, and 3) not
early-returning if the highest Z in the surrounding slot does not exceed to the pipette bounds at
that slot.

Closes RAUT-1046