automatically assign IP and expected volume moved (#15597)
<!--
Thanks for taking the time to open a pull request! Please make sure
you've read the "Opening Pull Requests" section of our Contributing
Guide:


https://github.com/Opentrons/opentrons/blob/edge/CONTRIBUTING.md#opening-pull-requests

To ensure your code is reviewed quickly and thoroughly, please fill out
the sections below to the best of your ability!
-->

# Overview

<!--
Added section to read a json file in the measurement folder containing
all robot IP's and expected volumes moved based on robot name.-->

# Test Plan

<!--
Tested steps 1, 2, and 3 with measurements. I do not believe more
testing is needed.
-->

# Changelog

<!--
- added json file reader for IP's and volumes
- added recovery if "Y" or "N" were not input after a measurement
-->

# Review requests

<!--
N/A
-->

# Risk assessment

<!--
This script is not relied on in any other script.
-->