feat(app, shared-data): implement CSV runtime parameter functionality on ODD
Update ODD ProtocolSetupParameters to accommodate file runtime paramter overrides. In the case that
the selected file already exists on the robot server, we simply add the existing file ID to
`runTimeParameterValues`. In the case that the file ID does not yet exist and we only have a
filePath from the selected file on a USB drive, we first send the filePath to /dataFiles to create a
file on the robot server, and send the returned ID with `runTimeParameterValues`