feat(app): CategorizedStepContent component (#15586)
This is a component that implements
https://www.figma.com/design/8dMeu8MuPfXoORtOV6cACO/Feature%3A-Error-Recovery-August-Release?node-id=4113-73992&t=oJBTLKrTkARxEUyD-4
(https://opentrons.atlassian.net/browse/EXEC-499). It's used in error
recovery.

~There are some open questions and todos around the step indices which
is why this is a draft. Specifically,~
- ~What's the sizing behavior of the step numbers in the desktop
version? As-implemented, they'll be at least 3 characters wide and then
all increase size in lockstep to the longest one~ that's fine
- ~I'm guessing at the font on desktop for the step indices~ I guessed
right, it was caption regular



~Also, I'm going to implement this in the currently-existing places that
use it before undrafting.~ conveniently there was just one

# to do:
- [x] ODD layout test
- [x] make sure nothing is more broken on desktop (you can't get to this
in desktop anyway)

Storybook:
https://s3-us-west-2.amazonaws.com/opentrons-components/exec-499-two-step-component/index.html?path=/story/app-molecules-interventionmodal-categorizedstepcontent--categorized-step

Closes EXEC-499