fix(app): odd: keep bullets on runningprotocol low
These little bullets that establish the ability to swipe left and right
need to be relatively above the rest of the runningprotocol screen yes
but they shouldn't be above anything else. So create an explicit
z-stacking context for the content part of running protocol.

I'm not quite sure why this is now necessary and wasn't previously, but
the rules around what creates an implicit stacking context are complex
so we probably changed something 8 components away that removed whatever
stacking context existed previously.