fix(app): odd: fix ER fundamental layout issues
I love Box. I don't know why. I always want to use it and every single
time it breaks something. This is one of those times

- Remove Box as the container for intervention modal because it makes
all of our common tools for layout not function
- Make the ODD containers 100% instead of explicit size everywhere
because if you are using the right kind of container they won't overflow
anyway
- 3am going back for more <Box />: inRecovery FooterButtons, add a small
dummy element to the button row when there is no "go back" button so the
primary button is still on the right