feat(shared-data): bump gripper model version and use new force polynomial table for new motor (#15413)
<!--
Thanks for taking the time to open a pull request! Please make sure
you've read the "Opening Pull Requests" section of our Contributing
Guide:


https://github.com/Opentrons/opentrons/blob/edge/CONTRIBUTING.md#opening-pull-requests

To ensure your code is reviewed quickly and thoroughly, please fill out
the sections below to the best of your ability!
-->

# Overview
The manufacturer is changing the brushes on the gripper motor, so we'll
need to use a new force polynomial for the new gripper models (v1.3). In
order to use the new force function, you will need to make sure the
serial number is in the format of `GRPV13xxxxx`.