refactor(app): TextOnlyButton
This is a new button that is in the app components. Its existence is a
bit tricky.

This is a kind of button that is just text of a particular color with
particular hover states - no background color, no border, etc. It's
meant to be the lowest-visual-priority interactable it can be.

We have a component that does this, for the ODD only. It's the
tertiaryLowLight state of SmallButton, which you can see in the
ODD/Atoms/Buttons tree in storybook. This component definitely wasn't
and isn't responsive to desktop styles.

Separately, we heavily used buttons of this style in the modern wizards:
PipetteWizard, GripperWizard, LPC. These wizards all define their
content bodies via GenericWizardTile, whose stories are under
App/Molecules/GenericWizardTile, and _that_ component had a completely
separate definition of this button style as a styled-component inline in
its implementation.

Now, we want to use a button of that style on desktop, but in Error
Recovery, which is a flow that doesn't use GenericWizardTile (because it
needs the border look&feel of InterventionModal). So the button needs to be
defined outside of GenericWizardTile, and also it needs to be
responsive. That means we could do one of the following:
1. Make all the ODD buttons responsive, somehow dealing with the case(s)
that have no semantic equivalent on desktop
2. Make just that one ODD button responsive, but for all of its possible
style types, see above
3. Make just that one ODD button in that one style responsive, but then
passing a different style as a prop would result in a weird render
4. Just make a new dang component

I picked 4. I don't really feel that great about it.