refactor(app): Refactor ER containers
The container was just the ContentWrapper and was implicitly only single
column; make versions for two columns and optional numbers of columns as
well, since those are layouts that are used throughout ER. Also handle
footer buttons, since those can be a little tricky (they should be
inside the content, or can be anyway, if SingleColumn but definitely
outside and spanning if TwoColumn).

This requires renaming the component instantiations in a whole bunch of
places, which is annoying, but such is life.