feat(app): ER: Desktop styles for SelectRecoveryOption
This implements the desktop styling for SelectRecoveryOption.
Specifically,

- On desktop this should be a twocolumn with failedstepnextstep, and on
ODD this should be a onecolumn with just the buttons; that means use the
fancy one or two columns wrapper
- The radio buttons need desktop styles, so use the fancy
RecoveryRadioGroup component

Unfortunately, since we're rendering FailedStepNextStep we need access
to the full analysis in tests which means we need to change the fixtures
which means some other components needed to handle having the fixtures changed.