feat(app, api-client, react-api-client): update POST interface for file parameters (#15682)
Previously, file-type parameters were included in the
`runTimeParameterValues` object used in posting to /protocols, /runs,
and /analyses. Now, a robot server refactor creates an expectation by
these endpoints for a separate `runTimeParameterFiles` object to be
posted for protocols including a file-type parameter.
`runTimeParameterValues` will be reserved for value-type parmeters. This
PR introduces a new utility to retrieve these parameters and return them
in their own object, and implements them in the required places across
api-client and react-api-client.