refactor(app): Desktop implementation for SelectRecoveryOption (#15691)
This PR got away from me a little bit (and I still have to test it - the
sum of these two is why it's in draft). It's probably best to view it
commit by commit.

Overall, the point is to implement the desktop styling for the
SelectRecoveryOption component of error recovery, here:
https://www.figma.com/design/8dMeu8MuPfXoORtOV6cACO/Feature%3A-Error-Recovery-August-Release?node-id=4829-78111&t=bTwOek0mZSA61ugm-4
. This is a little weird because it is exactly semantically equivalent
to the ODD panel, but has a very different styling - two columns instead
of one. It also uses the radio buttons that we use only in OT-2 tip
length calibration method selection, which are old and use cssmodules
and need a cleanup.

The approach here was to add some fundamental structure components to
InterventionModal that will render two columns on desktop (as the
TwoColumn structure, including min-size and wrap) and one column on the
ODD at full width. Then, we can build on top of that in ErrorRecovery,
with the big difference being that the ErrorRecovery component also
folds in the standard ER footer (or will - already had to find/replace a
bunch of stuff, going to move other components over to specifying their
footers through the wrapper as we update their styles).

There were also some incidental refactors. By commit, 
- 981e82802c06c1c5890fe2cb45ce24ec704a6775 : we have these utility
components for visualizing structure components; make them handle
sometimes being size limited and sometimes not being
- a7917e746138d314a0b76f4398fc0bcef5ec4877 : Add a wrapper around the
RadioGroup component. We're going to need to update these styles further
(that's a todo) but we don't want to, or I don't want to, have to mess
with cssmodules. Also I think @TamarZanzouri is touching this stuff at
the same time, so keep it isolated and droppable. Did I let the worms in
my brain drive and make a typescript wrapper for getting the change
events to take an inferred union of button values? Yes. Also note the
one change to the underlying component to specify IDs so that we can use
aria label linking in the passed components, which is both good practice
and allows tests to work later.
- e835f8604e5bcaf3430b67d21155562919a883d5 The error recovery component
for viewing failed and next commands was bundled with presenting text in
the left column, which is not what we want this panel to have, so split
it out.
- 699cb79b035aa157aca8544b9bddb9d5035eaedf The first fun one: Add a
component that can responsively present one or two columns, and by
"responsively" i mean abuse css mediaquery to only do it on touchscreen
and maintain the two-column responsive presentation through desktop
resizes. Best appreciated in storybook.
- 52150cbdcd24fa87c1bb754b6d1296444b4ed691 Simultaneous refactor and
feature - we had a RecoveryContentWrapper but it was really just a
single column, this isn't helpful, add a single and twocolumn version
(lots of find and replace since I changed the name) and then add a
OneOrTwoColumn on top of that. These wrappers also encompass rendering
footers because the footer needs to be in different places in one or two
column
- 00f38e8e4db80b1c16f2c4395b36c963a5bffa25 Use all that stuff to render
SelectRecoveryOption! Note that the tests now run on both presentations
and it all works because of that aria label stuff.

## Todo
- [x] Visual tests, at all (this is why it's a draft)