feat(robot-server): provide GET endpoints for /dataFiles (#15701)
# Overview

Closes AUTH-556.

This PR adds two GET endpoints to the `/dataFiles` route, one for
getting information on the data file (id, created at timestamp, file
name) and one for downloading the contents of the data file.

The one to get information about the dataFile is accessible via `GET
/dataFiles/{data_file_id}` and the one for downloading is `GET
/dataFiles/{data_file_id}/download`. The file is returned as the
MIME/media type `text/plain` for now. Both of these endpoint will return
with a status code of 200 for a success and 404 if the request file id
does not exist.

# Test Plan

A combination of unit and integration test coverage as well as tests on
robot via Postman.

# Changelog

- added `GET /dataFiles/{data_file_id}` to get information on that
specific data file
- added `GET /dataFiles/{data_file_id}/download` to download the data
file

# Review requests

Do we want to add some sort of format/media_type argument to the
download endpoint for a future where we have files we don't want to
label as `text/plain`? This could be done in a future PR, but if its
after a release we'd probably want to have it default to `text/plain`
for backwards compatability.

# Risk assessment

Low.