build: Sign windows builds via remote key storage (#15718)
Integrates code signing with a key stored in digicert one to app build
workflows.

There are a couple caveats:
- You can do this locally if you have a windows machine and if you have
the right accounts and permissions. Read: you basically can't do this
locally
- Digicert for some reason charges per signature. We sign a lot of
stuff. Therefore, we are only going to produce signed windows builds for
releases and if a dev really needs to by pushing a branch that has
"as-release" in it (in the same way we only do app builds if you push a
branch that has app-build in it - so building and signing both windows
apps would require a branch that has app-build-both-as-release in it)
- This just doesn't work at all with electron-builder, and they don't
seem to want to change things to fix it; specifically, you can only
configure e-b to pass along a key link and a password, and you basically
can't do that anymore. So we have to have a (thankfully simple) custom
sign script.

Closes RDEVOPS-128

## to leave draft
- [x] this produces a signed installer that passes smartscreen
- [x] this only does that on branches with the right kinds of names