fix(app): add fallback text to recovery success toasts (#15763)
Closes EXEC-624

If error recovery occurs for a step not in protocol analysis, or error recovery results in skipping to a command when the user is recovering from the last command of the run, there is no command text. As a fallback in these spots, let's just not show the full command text and just show the recovery action text. Note that this never occurs on the ODD, since we never show the full command text.