feat(api): Special error info for overpressure
We can create error-specific overloads for certain errors that have
well-defined error information, like in this case a retry location for
in-place errors.

This unfortunately means that the base class errorInfo very particularly
needs the type Mapping[str, object]. This is because python typing is
default-structural, so if you had this:

class A(TypedDict):
   key1: float

class B(TypedDict):
   key1: float
   key2: str

def f(m: Mapping[str, float]) -> None:
   pass

then if mypy tried to make typedicts match the Mapping, it would have to
allow B to match as well. Or so quoth the maintainers in
https://github.com/python/mypy/issues/8420 .

Anyway, it's an object.