feat(app,api): Move to right position before retrying in place command (#15776)
This PR adds something in between a nice feature and a hack for retrying
commands like `AspirateInPlace` in error recovery. The fundamental
problem is that these commands' parameters do not contain the location
where you're running them, and we know that we've moved the gantry since
they failed.

The fix is to add an optional `retryLocation` value to the error info of
the well-defined overpressure command (and subsequently to any other
well-defined error that will be raised from these commands), as in
2ae195619b6e19e3163528b05b3ba3e8300fac49 ; then use that location as the
target of a MoveToCoordinates before we actually retry the target
command if necessary, as we do in
9183301e27d351ad4a2b560ad29a52c73ed99c36

## Testing
- [x] Pass tests, mostly. I wanted to keep this separate from actually
implementing the errors in those commands just to have a nicer commit
log, so you can't really test it right now. I figured I'd follow up with
a PR that implements it in those commands and apply any needed fixes
there.