feat(api): Handle overpressures in AspirateInPlace
Use DefinedErrorData for pipette overpressures in AspirateInPlace.

This is mostly similar to adding the same handling for Aspirate and
Dispense, but the big difference is that these commands are not
associated with wells or with points via their params.

We get the current point by asking the hardware where it is; it's
important to do this rather than getting it from the state because
someone could conceivably have called AspirateInPlace as the first
command after a move command with a different pipette, and we wouldn't
have the point in state.

We don't get the current well at all. That's because this command isn't
associated with a well, so there's not really anything to change.

Closes EXEC-496