fix(app): prevent eternal intervention purgatory (#15788)
If you click on a run that has an intervention modal due to a pause, move labware, etc., you shouldn't have to stare at that run until you close or your app, deal with the robot, or contemplate your navigation actions for all eternity. Navigating away should be a legitimate option.