feat(ci): analyses snapshot skip pr on fail label (#15792)
## Overview

- ~~Keep making the analyses snapshot test workflow more flexible
allowing a label on the PR to prevent failure PRs from opening. Adding
the label `no-analyses-snapshot-pr` on your PR will cause the workflow
to skip opening a PR for snapshot updates if the snapshot tests fail.~~
- Added a step to diff the snapshots in the PR target branch against the
PR branch. No action is taken based on the outcome. Purely
informational. Will this information be noticed and used as a reminder
to synchronize with the target branch?

### Updates from feedback

- default `OPEN_PR_ON_FAILURE` to false 
- as is the case currently if the analyses snapshot test fails you will
see a red X on that test and you can read the logs or download and view
the HTML test report.
- if you want to open a PR into your branch that would heal the
snapshots, add `gen-analyses-snapshot-pr` label and a PR will get opened
- create a specific step to handle PRs for the overnight scheduled test
of edge.

### Risk

No risk or change to application code, CI only