feat(api): Add HCPAPI to update from encoders
We have encoders, at least on the flex; we now need to use them in
actually reasonably public APIs.

Add the minimal API to take advantage of the encoders with a wrapper
around the already-existing backend update_position_estimators function
that takes a list of axes to update the estimators for. This will cause
these axes to update their open-loop estimators from the encoders. We
don't want to do this frequently because the encoders have a coarser
resolution than the internal position accumulators, so this will be a
net accuracy loss; these axes should be homed before being used for
something important. However, the encoder is plenty accurate for blowing
out or dropping tips to make the pipette safe to home, and that's what
it's for.

Also slightly refactor the hardware protocols so this and the other
encoder stuff is in its own protocol facet.