feat(api): add unsafe/dropTipInPlace
This is a special command that (on Flex) will make the machine reload
its plunger position from its encoder and then drop tip. This makes this
command always runnable even if there's no position known for the
plunger axis, as there wouldn't be after an error during a liquid
handling command. This lets us call it unconditionally during the drop
tip wizard.