feat(api): add unsafe/UpdatePositionEstimators
Adds a command that will update all specified axis position estimators from their
encoders. This is something that a client can use when it knows it's
about to be moving multiple axes. As with the other unsafe commands,
it's best to home after doing this before doing something that requires accuracy.