refactor(robot-server): move run time params to their own DB table (#15650)
Closes AUTH-527

# Overview

This PR moves primitive RTPs to their own table and refactors the
analysis-generation process. This entire refactor makes it easier to add
on CSV-parameter handling.

Architecture and database changes:

- Created a new table to store all primitive (non-CSV) run time
parameter values used in the analysis.
- Previously, we used to compare only the RTP values provided by the
client with the full set of RTP values used in the last analysis.
Because the values provided by the client cannot be assumed to be the
full set of values of run time parameters in a protocol (parameters that
don't have a client-specified value get assigned a default one), we had
to make additional comparison of previous RTP values and default values.
- This PR simplifies that comparison by extracting the full set of RTP
values for the current analysis beforehand and then comparing it with
the full set of previous RTP values.
- This extraction of full RTP values earlier in the process meant that
we needed to re-architect the flow of analysis.

# Test Plan

- Since this is only a refactor, existing integration and snapshot tests
passing should be a good enough check for whether this is accidentally
changing anything it shouldn't be
- This changes the DB by removing a column from a table in schema 4 so
we should make sure that a robot on DB v4, that has protocols (both RTP
& non-RTP), gets migrated to v5 successfully and there are no issues
handling their analyses.

# Changelog

- added the new primitive RTP table and removed the 'rtp values and
default' column from analysis table
- updated DB v4 -> v5 migration
- refactored `AnalysesManager`, `ProtocolAnalyzer`, `AnalysisStore` and
protocols `router` to use the new flow as shown above.
- updated tests 

# Review requests

- @SyntaxColoring and I had discussed most of these changes pretty
extensively. I was initially hesitant towards moving the RTP validation
and extraction earlier in the process because of the amount of
refactoring that would result in, but as i was working towards adding
CSV parameter handling, the early validation method deemed quite useful,
especially since CSV parameters will be relying on external files to set
their 'values'. (so thanks for the suggestion Max!)
- Does the new flow make sense? Anything that isn't clear?

# Risk assessment

Medium-high. It changes quite a critical part of the protocol upload and
analysis process. But the fact that all integration tests and analyses
snapshots are passing makes me feel quite confident that the existing
behavior of these endpoints has not changed by this work.