fix(app): allow completed runs when starting protocol
This means that people don't have to make sure to explicitly close the
run context by clicking the x button in the post-run flow before picking
a new robot to run.

Use the current run's completed at timestamp to decide whether or not
the run is complete for the purposes of idle. This requires fetching the
current run, which we do by ID.

Closes EXEC-519