fix(app): Allow starting protocols on robots with completed runs (#15839)
When you click the button to start protocol setup, we try to filter to
only robots that are currently capable of starting a protocol. We
implemented that as
- You can connect to the robot
- The robot doesn't have a current run

But, robots don't actually get rid of their current run until a user
either clicks the little X button in the run screen on the desktop or
goes back to the main screen of the ODD. That doesn't necessarily align
with a user's view of "this protocol is done", and so people would think
that the robot just disappeared from that little slideout.

To fix this, we should also consider robots that have a completed
protocol to be done. This is a small conceptual change and unfortunately
a huge pain to implement because data about whether a run is done has to
come from getting the state data for that run.

## Testing
- [x] The setup protocol slideout should show robots that have
completed, failed, or cancelled active runs
   - [x] If you start setup, it should work
- [x] The setup protocol slideout should show robots that do not have an
active run
- [x] The setup protocol slideout should not show robots that have an
active run that is not completed, failed, or canceled

Closes EXEC-519