fix(app): fix "return to dashboard" not returning to dashboard
After clicking "return to dashboard" on the ODD, the run context is closed, and then a navigate to
dashboard is issued. However, the top level redirects logic keeps the ODD on the runs summary page
as long as there is an active run record, which is always the case when the network request hasn't
yet completed. By the time the user clicks "return to dashboard" a second time, the run has closed.
Unfortunately, throwing this navigate behind an onSuccess callback is insufficient, because the top
level redirects hook uses a different endpoint to get the run info. Instead, let's eagerly clear the
query cache on the ODD for the run record. Any future refetches of that run record simply repopulate
the cache. It's worth revisiting our approach to top level redirects in the future, but this
solution is adequate for now.