feat(app): Implement `hasEverEnteredErrorRecovery` (#15876)
Closes EXEC-636 and EXEC-637 and RQA-2900

Because ER always requires the users to go through drop tip flows, it's redundant and mildly annoying for users to see a second set of now irrelevant CTAs. Now, we only show the CTAs at the end of the run if the run did not enter error recovery (and the conditions for detecting potential tip attachment occur).